対応していないコーデックは skip するようにする
